지난밥/lastmeal/requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
asgiref==3.2.10
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
#brotlipy==0.7.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
drf-yasg==1.17.1
gast==0.3.3
google-api-core==1.7.0
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-cloud-vision==0.35.2
google-pasta==0.2.0
googleapis-common-protos==1.5.5
grpcio==1.32.0
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
inflection==0.5.1
ipython-genutils==0.2.0
itypes==1.2.0
Jinja2==2.11.2
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
multidict==4.7.6
mysql==0.0.2
mysqlclient==2.0.1
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.4
parso==0.7.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyreadline==2.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
#pywin32==227
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.4.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tb-nightly==1.14.0a20190603
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.0
zipp==3.1.0
